+++ libs/compiler/preprocessor/file.c
@@ -20,7 +20,7 @@
 #include "context_var.h"
 #include "preprocessor_error.h"
 #include "preprocessor_utils.h"
-#include "uniprinter2.h"
+#include "uniprinter.h"
 #include <limits.h>
 #include <math.h>
 #include <stdio.h>
