--- libs/utils/uniprinter_file.c
@@ -1,35 +1,0 @@
-/*
- *	Copyright 2019 Andrey Terekhov
- *
- *	Licensed under the Apache License, Version 2.0 (the "License");
- *	you may not use this file except in compliance with the License.
- *	You may obtain a copy of the License at
- *
- *		http://www.apache.org/licenses/LICENSE-2.0
- *
- *	Unless required by applicable law or agreed to in writing, software
- *	distributed under the License is distributed on an "AS IS" BASIS,
- *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *	See the License for the specific language governing permissions and
- *	limitations under the License.
- */
-
-#include "uniprinter.h"
-#include <limits.h>
-#include <math.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-
-/** File-based printer */
-int printer_file_fprintf(universal_printer_options *opts, const char *fmt, va_list args)
-{
-	int ret;
-
-	ret = vfprintf(opts->output, fmt, args);
-
-	return ret;
-}
-
-printer_desc printer_file = { IO_SOURCE_FILE, printer_file_fprintf };
