+++ libs/utils/uniscanner.h
@@ -1,5 +1,5 @@
 /*
- *	Copyright 2019 Andrey Terekhov
+ *	Copyright 2020 Andrey Terekhov, Victor Y. Fadeev, Dmitrii Davladov
  *
  *	Licensed under the Apache License, Version 2.0 (the "License");
  *	you may not use this file except in compliance with the License.
@@ -16,114 +16,35 @@
 
 #pragma once
 
+#include <stdio.h>
 #include "io.h"
 #include "dll.h"
-#include <stdarg.h>
-#include <stdbool.h>
-#include <stdio.h>
 
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-/** Options of universal scanner */
-typedef struct universal_scanner_options
-{
-	ruc_io_source source; /** Data source */
-	FILE *input;		  /** Input file (in case of file source) */
-	char *ptr;			  /** Data pointer (in case of memory source) */
-	int pos;			  /** Position to data within @p ptr */
-	void *opaque;		  /** Scanner's opaque data */
-} universal_scanner_options;
+/**
+ *	Universal scanf-like function
+ *
+ *	@param	io			Universal io structure
+ *	@param	format		String format
+ *
+ *	@return	Return scanf-like value
+ */
+EXPORTED int uni_scanf(universal_io *const io, const char *const format, ...)
+	__attribute__((format(scanf, 2, 3)))
+	__attribute__((nonnull(2)));
 
 /**
- *	Prototype for a function getting a next symbol from the context
+ *	Universal function for scanning UTF-8 characters
  *
- *	@param	opts	Scanner options
+ *	@param	io			Universal io structure
  *
- *	@return	Read symbol
+ *	@return	UTF-8 character
  */
-typedef int (*io_getnext_t)(universal_scanner_options *opts);
-
-/**
- *	Prototype for a scanf-like function
- *
- *	@param	opts	Scanner options
- *	@param	fmt		Format string
- *	@param	args	List of arguments
- *
- *	@return	scanf()-like result
- */
-typedef int (*io_scanf_t)(universal_scanner_options *opts, const char *fmt, va_list args);
-
-/** Scanner description */
-typedef struct scanner_desc
-{
-	ruc_io_source source; /** Data source supported by scanner */
-	io_getnext_t getnext; /** The pointer to a function scanning the input from the context */
-	io_scanf_t scanf;	  /** the pointer to a scanf-like function */
-} scanner_desc;
-
-
-/**
- *	Initialize scanner
- *
- *	@param	opts	Scanner context
- */
-EXPORTED void scanner_init(universal_scanner_options *opts);
-
-/**
- *	Deinitialize scanner
- *
- *	@param	opts	Scanner context
- */
-EXPORTED void scanner_deinit(universal_scanner_options *opts);
-
-/**
- *	Close current scanner context
- *
- *	@param	opts	Scanner context
- */
-EXPORTED void scanner_close(universal_scanner_options *opts);
-
-/**
- *	Get a next symbol from input stream
- *
- *	@param	opts	Scanner context
- *
- *	@return	Read symbol
- */
-EXPORTED int scanner_getnext(universal_scanner_options *opts);
-
-/**
- *	scanf() for the uniscanner stream
- *
- *	@param	opts	Scanner context
- *
- *	@return	scanf()-like return value
- */
-EXPORTED int scanner_scanf(universal_scanner_options *opts, const char *fmt, ...);
-
-/**
- *	Attach file to scanner
- *
- *	@param	opts	Universal scanner options
- *	@param	file	Target file
- *
- *	@return	@c true on success, @c false on failure
- */
-EXPORTED bool scanner_attach_file(universal_scanner_options *opts, FILE *file);
-
-/**
- *	Attach buffer to scanner
- *
- *	@param	opts	Universal scanner options
- *	@param	ptr		Target buffer
- *
- *	@return	@c true on success, @c false on failure
- */
-EXPORTED bool scanner_attach_buffer(universal_scanner_options *opts, const char *ptr);
+EXPORTED char32_t uni_scan_char(universal_io *const io);
 
 #ifdef __cplusplus
 } /* extern "C" */
